Refactor: Move MyReads state to BooksApp component
Move MyReads state and state manipulation to BooksApp component so that there is only one source of truth for books state. When we search for books it also returns already shelved books, i'll need to access the my books to filter out the already shevled books from search results. For that the books state needs to be in a higher order component so that both search and list books components can access it.